rename lib to cognitiveSteps
