+++ packages/soul-engine-cli/template/package.json
@@ -15,6 +15,6 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@opensouls/engine": ">=0.1.4"
+    "@opensouls/engine": ">=0.1.7"
   }
 }
