+++ packages/soul-engine-cli/template/soul/cognitiveSteps/externalDialog.ts
