+++ packages/soul-engine-cli/template/soul/cognitiveSteps/internalMonologue.ts
