+++ packages/soul-engine-cli/template/soul/cognitiveSteps/mentalQuery.ts
