+++ packages/soul-engine-cli/template/soul/initialProcess.ts
@@ -1,8 +1,8 @@
 
 import { MentalProcess, useActions, useProcessManager } from "@opensouls/engine";
 import shouts from "./mentalProcesses/shouts.js";
-import externalDialog from "./lib/externalDialog.js";
-import mentalQuery from "./lib/mentalQuery.js";
+import externalDialog from "./cognitiveSteps/externalDialog.js";
+import mentalQuery from "./cognitiveSteps/mentalQuery.js";
 
 const gainsTrustWithTheUser: MentalProcess = async ({ workingMemory }) => {
   const { speak, log  } = useActions()
