+++ packages/soul-engine-cli/template/soul/subprocesses/learnsAboutTheUser.ts
@@ -1,7 +1,7 @@
 
 import { ChatMessageRoleEnum, MentalProcess, WorkingMemory, createCognitiveStep, indentNicely, useActions, useProcessMemory } from "@opensouls/engine";
-import mentalQuery from "../lib/mentalQuery.js";
-import internalMonologue from "../lib/internalMonologue.js";
+import mentalQuery from "../cognitiveSteps/mentalQuery.js";
+import internalMonologue from "../cognitiveSteps/internalMonologue.js";
 
 const userNotes = createCognitiveStep(() => {
   return {
